canEmg送ると止まるバグ修正｡たぶんreadStatまわりがバグってる(どっかで消しちゃってるので過電流保護がいまいち働かないっぽい

+++ CanDCMD_ver1.0/user/user_app/inc/canVoltage.h
@@ -8,12 +8,12 @@
 class CanVoltage{
 public :
 	void setup(CAN &can,uint16_t number,uint16_t interval);
-	int32_t read();
+	float read();
 	void interrupt();
 
 	uint64_t lastReceiveTime();
 	uint64_t receiveTime;
-	int32_t value;
+	float value;
 	uint16_t number;
 
 private:
